IN-89 refactor: change code after review
