+++ src/feature/auth/index.ts
@@ -12,6 +12,6 @@ export { Congratulations } from './ui/Congratulations/Congratulations'
 export { CreateNewPasswordForm } from './ui/CreateNewPasswordForm/CreateNewPasswordForm'
 export { EmailVerificationBlock } from './ui/EmailVerification/EmailVerification'
 export { ForgotPasswordForm } from './ui/ForgotPasswordForm/ForgotPasswordForm'
-export { LogOut } from './ui/LogOut/LogOut'
+export { LogoutDialog } from './ui/LogoutDialog/LogoutDialog'
 export { SignInForm } from './ui/SignInForm/SignInForm'
 export { SignUpForm } from './ui/SignUpForm/SignUpForm'
