+++ src/feature/auth/ui/LogoutDialog/LogoutDialog.module.scss
