+++ src/widgets/dialogs/index.ts
@@ -1,2 +1,2 @@
+export { ConfirmDialog } from './ui/ConfirmDialog/ConfirmDialog'
 export { DialogEmailSent } from './ui/DialogEmailSent/DialogEmailSent'
-export { DialogLogOut } from './ui/DialogLogOut/DialogLogOut'
