+++ src/widgets/dialogs/ui/ConfirmDialog/ConfirmDialog.module.scss
